refactor: move searchParams utils inside mySearchParams hook
